refactor: reusing useColumns
