+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingListsCommon/useColumns/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { useColumns } from './useColumns';
