+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingListsCommon/useDefaultColumnConfiguration/index.js
