+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingListsOffline/ColumnSetup/EventWorkingListsOfflineColumnSetup.component.js
@@ -1,6 +1,6 @@
 // @flow
 import React, { useMemo } from 'react';
-import { useDefaultColumnConfig } from '../../EventWorkingListsCommon';
+import { useDefaultColumnConfig, useColumns } from '../../EventWorkingListsCommon';
 import { EventWorkingListsOfflineDataSourceSetup } from '../DataSourceSetup';
 import type { Props } from './eventWorkingListsOfflineColumnSetup.types';
 
@@ -10,24 +10,7 @@ export const EventWorkingListsOfflineColumnSetup = ({
     ...passOnProps
 }: Props) => {
     const defaultColumns = useDefaultColumnConfig(program);
-
-    const defaultColumnsAsObject = useMemo(() =>
-        defaultColumns
-            .reduce((acc, column) => ({ ...acc, [column.id]: column }), {}),
-    [defaultColumns]);
-
-    const columns = useMemo(() => {
-        if (!customColumnOrder) {
-            return defaultColumns;
-        }
-
-        return customColumnOrder
-            .map(({ id, visible }) => ({
-                ...defaultColumnsAsObject[id],
-                visible,
-            }));
-    }, [customColumnOrder, defaultColumns, defaultColumnsAsObject]);
-
+    const columns = useColumns(customColumnOrder, defaultColumns);
     const visibleColumns = useMemo(() => columns.filter(column => column.visible), [columns]);
 
     return (
